backend/requirements.txt
torch>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
onnx>=1.15.0
# onnxruntime>=1.16.0  # Optional - only needed for model validation, not training
Pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0
scikit-learn>=1.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
tqdm
pytest
moderngl>=5.8.0
glfw>=2.6.0